# perf: do not store the whole file in memory but check while reading