Rework the ContainsHeader to check the header while reading the file
instead of storing the whole file in memory.
We also compare bytes to avoid string allocation.

Benchmarks results:

```
name        old time/op    new time/op    delta
Run/dot-16    1.02ms ±15%    0.70ms ±10%  -31.21%  (p=0.000 n=10+10)

name        old alloc/op   new alloc/op   delta
Run/dot-16     241kB ± 0%      95kB ± 0%  -60.74%  (p=0.000 n=10+10)

name        old allocs/op  new allocs/op  delta
Run/dot-16     2.48k ± 0%     0.84k ± 0%  -66.16%  (p=0.000 n=10+10)
```